requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-enumfields==0.10.1
django-environ==0.4.4
django-filter==1.1.0
django-js-asset==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
future==0.16.0
gunicorn==19.8.1
idna==2.6
MySQL-python==1.2.5
Pillow==5.1.0
PyJWT==1.6.1
python-dotenv==0.8.2
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
